+++ rpc/src/main/java/com/walmartlabs/concord/rpc/CommandHandler.java
@@ -1,0 +1,8 @@
+package com.walmartlabs.concord.rpc;
+
+public interface CommandHandler {
+
+    void onCommand(Command cmd);
+
+    void onError(Throwable t);
+}