+++ rpc/src/main/java/com/walmartlabs/concord/rpc/CommandQueue.java
@@ -2,5 +2,5 @@ package com.walmartlabs.concord.rpc;
 
 public interface CommandQueue {
 
-    Command take() throws ClientException;
+    void stream(CommandHandler handler) throws ClientException;
 }