+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log
 
+## [Unreleased]
+
+### Added
+
+- concord-server: automatically remove orphaned data.
+
+### Changed
+
+- ansible: throw an exception if a private key file was not found.
+
+
+
 ## [0.43.0] - 2017-10-05
 
 ### Added