+++ CHANGELOG.md
@@ -9,8 +9,8 @@
 
 ### Changed
 
-- concord-rpc: fix receiving of commands by the agent when the server
-is restarted;
+- concord-rpc: fix dispatching of agent commands when the server is
+restarted;
 - ansible: throw an exception if a private key file was not found.
 
 ### Breaking